p4i integration Integrate 273786 from 4.3 to main: Fixes:    compiler warning AutoTest: (the warning one) http://kramer/makelog___compile_qt____desert__4_3_2__linux-g++.html
[git-p4: depot-paths = "//depot/qt/main/": change = 273787]

+++ src/gui/kernel/qwidget_x11.cpp
@@ -331,7 +331,7 @@ static QVector<Atom> getNetWmState(QWidget *w)
                                returnValue.size(), False, XA_ATOM, &actualType, &actualFormat,
                                &propertyLength, &bytesLeft, &propertyData) != Success) {
             returnValue.clear();
-        } else if (propertyLength != returnValue.size()) {
+        } else if (propertyLength != (ulong)returnValue.size()) {
             returnValue.resize(propertyLength);
         }
